Develop(helpers.test): Wrote test for fetchNewAPIData()
